requirements/base.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1.2
django-filter==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
django-widget-tweaks==1.4.3
idna==2.7
Pillow==5.3.0
pytz==2018.5
requests==2.19.1
six==1.12.0
urllib3==1.23
python-decouple==3.1
virtualenvwrapper==4.8.2
